+++ nexus_constructor/model/__init__.py
@@ -1,3 +1,4 @@
 from .group import Group  # noqa: F401
 from .entry import Entry  # noqa: F401
 from .component import Component  # noqa: F401
+from .group_container import GroupContainer  # noqa: F401
