+++ nexus_constructor/model/group.py
@@ -179,16 +179,3 @@ def name_not_in_excludelist(child: Any):
     if hasattr(child, CommonKeys.NAME) and child.name in CHILD_EXCLUDELIST:
         return False
     return True
-
-
-class GroupContainer:
-    def __init__(self, group: Group):
-        self._group = group
-
-    @property
-    def group(self) -> Group:
-        return self._group
-
-    @group.setter
-    def group(self, new_group: Group):
-        self._group = new_group
