+++ nexus_constructor/widgets/__init__.py
@@ -1,5 +1,7 @@
 from nexus_constructor.widgets.group_name_edit import GroupNameEdit  # noqa: F401
 from nexus_constructor.widgets.description_edit import DescriptionEdit  # noqa: F401
 from nexus_constructor.widgets.dropdown_list import DropDownList  # noqa: F401
-from nexus_constructor.widgets.class_dropdown_list import ClassDropDownList  # noqa: F401
+from nexus_constructor.widgets.class_dropdown_list import (  # noqa: F401
+    ClassDropDownList,
+)
 from nexus_constructor.widgets.custom_dialog import CustomDialog  # noqa: F401
